# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MK Store Backend API",
    "version": "1.0.0",
    "description": "E-commerce application backend API documentation with sample requests and responses",
    "contact": {
      "name": "API Support",
      "email": "support@mkstore.com"
    }
  },
  "servers": [
    {
      "url": "",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoints"
      }
    },
...
  },
...
  "paths": {
...
    "/api/save-product": {
      "post": {
        "summary": "Create a new product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "categoryId",
                  "vendorId",
                  "branchId",
                  "variants"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Wireless Headphones",
                    "description": "Product title (required)"
                  },
                  "categoryId": {
                    "type": "integer",
                    "example": 1,
                    "description": "Category ID (required)"
                  },
                  "subCategoryId": {
                    "type": "integer",
                    "example": 1,
                    "description": "Subcategory ID (optional)"
                  },
                  "brandId": {
                    "type": "integer",
                    "nullable": true,
                    "example": 1,
                    "description": "Brand ID (optional, can be null)"
                  },
                  "vendorId": {
                    "type": "integer",
                    "example": 1,
                    "description": "Vendor ID (required)"
                  },
                  "branchId": {
                    "type": "integer",
                    "example": 1,
                    "description": "Branch ID (required)"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ],
                    "example": "ACTIVE",
                    "description": "Product status (optional, defaults to ACTIVE)"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Multiple product image files (optional, max 3 images per product)"
                  },
                  "variants": {
                    "type": "string",
                    "description": "JSON string array of variant objects (REQUIRED - at least one variant). Each variant must include: variantName (required), description (optional), nutritional (optional), price (required), sellingPrice (required), quantity (required), itemsPerUnit (optional), units (optional), itemQuantity (optional), itemUnit (optional), expiryDate (required), status (optional), comboDiscounts (optional array). Each comboDiscount object includes: comboQuantity (required), discountType (required: PERCENT or OFFER), discountValue (required), startDate (required), endDate (required), status (optional)",
                    "example": "[{\"variantName\":\"500g\",\"description\":\"Fresh product\",\"nutritional\":\"Calories: 100\",\"price\":500,\"sellingPrice\":450,\"quantity\":100,\"itemsPerUnit\":2,\"units\":\"PCS\",\"expiryDate\":\"2024-12-31\",\"status\":\"ACTIVE\",\"comboDiscounts\":[{\"comboQuantity\":2,\"discountType\":\"PERCENT\",\"discountValue\":10,\"startDate\":\"2024-01-01\",\"endDate\":\"2024-12-31\",\"status\":\"ACTIVE\"}]}]"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Product created successfully"
                    },
                    "doc": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "title": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/get-product": {
      "post": {
        "summary": "Get products with pagination and filters",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pageSize": {
                    "type": "integer",
                    "enum": [
                      1,
                      5,
                      10,
                      20,
                      30,
                      40,
                      50,
                      100,
                      500
                    ],
                    "default": 10,
                    "description": "Number of results per page"
                  },
                  "pageNumber": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1,
                    "description": "Page number"
                  },
                  "filters": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "eq": {
                          "type": "string"
                        },
                        "in": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "neq": {
                          "type": "string"
                        },
                        "gt": {
                          "type": "string"
                        },
                        "gte": {
                          "type": "string"
                        },
                        "lt": {
                          "type": "string"
                        },
                        "lte": {
                          "type": "string"
                        },
                        "like": {
                          "type": "string"
                        },
                        "iLike": {
                          "type": "string"
                        }
                      }
                    },
                    "description": "Array of filter objects"
                  },
                  "sorting": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "direction": {
                          "type": "string",
                          "enum": [
                            "ASC",
                            "DESC"
                          ]
                        }
                      }
                    },
                    "description": "Array of sorting objects"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "doc": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "title": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "integer",
                      "example": 50
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/get-products-by-category": {
      "post": {
        "summary": "Get products grouped by category",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pageSize": {
                    "type": "integer",
                    "enum": [
                      1,
                      5,
                      10,
                      20,
                      30,
                      40,
                      50,
                      100,
                      500
                    ],
                    "default": 10,
                    "description": "Number of results per page"
                  },
                  "pageNumber": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1,
                    "description": "Page number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products grouped by category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "doc": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "title": {
                              "type": "string"
                            },
                            "price": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}